Studies Comparing Medications on Post-Operative Nausea and Vomiting

Outcome data from randomised controlled trials comparing anti-emetic medications (eg ondansetron, dexamethasone) against comparator anti-emetics or placebo in PONV outcomes for patients undergoing surgery.

Usage

binary_outcome

Format

A data frame with 12 rows and 18 variables:

Author

Study author

Year

Year study published

Subgroup

Surgery subgroup

Control

Control intervention

Intervention

Intervention medication

N_Total

Total number of patients in the study

N_Control

Number of patients in the control group

N_Intervention

Number of patients in the intervention group

Outcome_Control_Yes

Number of patients in the control group with the outcome (PONV)

Outcome_Control_No

Number of patients in the control group without the outcome (PONV)

Outcome_Intervention_Yes

Number of patients in the intervention group with the outcome (PONV)

Outcome_Intervention_No

Number of patients in the intervetion group without the outcome (PONV)

D1

Domain 1 of the Risk of Bias 2.0 tool

D2

Domain 2 of the Risk of Bias 2.0 tool

D3

Domain 3 of the Risk of Bias 2.0 tool

D4

Domain 4 of the Risk of Bias 2.0 tool

D5

Domain 5 of the Risk of Bias 2.0 tool

Overall

Overall risk of bias